# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication API",
        "description": "API endpoints for user authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get all games",
                "description": "List all games",
                "operationId": "e333e2e46e6c7bd4c254fe379b7389dd",
                "responses": {
                    "200": {
                        "description": "List of games",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "home_team": {
                                                "type": "object"
                                            },
                                            "away_team": {
                                                "type": "object"
                                            },
                                            "home_team_total_score": {
                                                "type": "integer"
                                            },
                                            "away_team_total_score": {
                                                "type": "integer"
                                            },
                                            "stats": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Create a new game",
                "description": "Create a new game",
                "operationId": "496a1db2ab2b344c40cad9e98d332d9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "home_team_id",
                                    "away_team_id"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "home_team_id": {
                                        "type": "integer"
                                    },
                                    "away_team_id": {
                                        "type": "integer"
                                    },
                                    "home_team_total_score": {
                                        "type": "integer"
                                    },
                                    "away_team_total_score": {
                                        "type": "integer"
                                    },
                                    "home_team_first_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_first_quarter_score": {
                                        "type": "integer"
                                    },
                                    "home_team_second_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_second_quarter_score": {
                                        "type": "integer"
                                    },
                                    "home_team_third_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_third_quarter_score": {
                                        "type": "integer"
                                    },
                                    "home_team_fourth_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_fourth_quarter_score": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "game": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get game details",
                "description": "Get a specific game",
                "operationId": "48fb06317bf9e8f10c22dd892ff7425e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "home_team": {
                                            "type": "object"
                                        },
                                        "away_team": {
                                            "type": "object"
                                        },
                                        "home_team_total_score": {
                                            "type": "integer"
                                        },
                                        "away_team_total_score": {
                                            "type": "integer"
                                        },
                                        "stats": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Games"
                ],
                "summary": "Update game details",
                "description": "Update a game",
                "operationId": "273f111abbd78231f5f9c48d4fca792c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "home_team_id",
                                    "away_team_id"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "home_team_id": {
                                        "type": "integer"
                                    },
                                    "away_team_id": {
                                        "type": "integer"
                                    },
                                    "home_team_total_score": {
                                        "type": "integer"
                                    },
                                    "away_team_total_score": {
                                        "type": "integer"
                                    },
                                    "home_team_first_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_first_quarter_score": {
                                        "type": "integer"
                                    },
                                    "home_team_second_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_second_quarter_score": {
                                        "type": "integer"
                                    },
                                    "home_team_third_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_third_quarter_score": {
                                        "type": "integer"
                                    },
                                    "home_team_fourth_quarter_score": {
                                        "type": "integer"
                                    },
                                    "away_team_fourth_quarter_score": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "game": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Delete a game",
                "description": "Delete a game",
                "operationId": "c5bffb910bb1804bac14c04327dad425",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}